Another test fix. Resolve namespace conflict by renaming
